change grabcut to use double precision for bgdModel and dgdModel as per opencv svn
git-svn-id: https://emgucv.svn.sourceforge.net/svnroot/emgucv/trunk@963 d7f09016-e345-0410-b530-edf29a71df78

+++ Emgu.CV.Test/AutoTestVarious.cs
@@ -1518,8 +1518,8 @@ namespace Emgu.CV.Test
          Image<Bgr, Byte> img = new Image<Bgr, byte>("airplane.jpg");
 
          Rectangle rect = new Rectangle(new Point(24, 126), new Size(483, 294));
-         Matrix<Single> bgdModel = new Matrix<float>(1, 13 * 5);
-         Matrix<Single> fgdModel = new Matrix<float>(1, 13 * 5);
+         Matrix<double> bgdModel = new Matrix<double>(1, 13 * 5);
+         Matrix<double> fgdModel = new Matrix<double>(1, 13 * 5);
          Image<Gray, byte> mask = new Image<Gray, byte>(img.Size);
 
          CvInvoke.CvGrabCut(img, mask, ref rect, bgdModel, fgdModel, 0, Emgu.CV.CvEnum.GRABCUT_INIT_TYPE.INIT_WITH_RECT);
